added more detailed docs to the Grid component
