+++ src/foundations/foundations.stories.mdx
@@ -15,7 +15,7 @@ import colorImg from '../assets/images/foundation_overview_color.png';
 import typographyImg from '../assets/images/foundation_overview_typography.png';
 import shapeImg from '../assets/images/foundation_overview_shape.png';
 import iconsImg from '../assets/images/foundation_overview_icons.png';
-import spacing from './layout/Spacing';
+import spacing from './layout/Grid.spacing';
 
 <Meta title="Foundations/Overview" />
