+++ src/foundations/layout/Grid.spacing.ts
@@ -1,22 +1,6 @@
-type TSpacingTokens = 0 | 25 | 50 | 75 | 100 | 150 | 200 | 250 | 300 | 350 | 400 | 450 | 500;
+import { TSpacing, TSpacingSides, TSpacingTokens, TSpacingTokensTRBL } from './Grid.types';
 
-type TSpacingSides = 'top' | 'right' | 'bottom' | 'left';
-
-type TSpacingTokensTRBL = {
-    [key in TSpacingSides]?: TSpacingTokens;
-};
-
-type TSpacing = {
-    spacing: (TSpacingTokens | null)[];
-    all: (spacingToken: TSpacingTokens) => TSpacing;
-    trbl: (spacingTokensTRBL: TSpacingTokensTRBL) => TSpacing;
-    only: (side: TSpacingSides, spacingToken: TSpacingTokens) => TSpacing;
-    vertical: (spacingToken: TSpacingTokens) => TSpacing;
-    horizontal: (spacingToken: TSpacingTokens) => TSpacing;
-    parseSpacing: () => string;
-};
-
-const Spacing = (): TSpacing => ({
+const GridSpacing = (): TSpacing => ({
     spacing: new Array(4).fill(null),
     all(spacingToken: TSpacingTokens): TSpacing {
         this.spacing.fill(spacingToken);
@@ -74,5 +58,4 @@ const Spacing = (): TSpacing => ({
     },
 });
 
-export type { TSpacing, TSpacingSides, TSpacingTokens };
-export default Spacing;
+export default GridSpacing;
