+++ src/foundations/layout/Grid.tsx
@@ -1,31 +1,13 @@
 import React from 'react';
 
+import { PGrid } from './Grid.props';
 import StyledGrid from './Grid.styles';
-import { TSpacing } from './Spacing';
-
-type PGrid = {
-    row?: boolean;
-    flex?: number;
-    component?: 'div' | 'span' | 'section' | 'aside';
-    alignment?: 'start' | 'center' | 'end' | 'stretch';
-    justify?:
-        | 'start'
-        | 'center'
-        | 'end'
-        | 'stretch'
-        | 'space-around'
-        | 'space-between'
-        | 'space-evenly';
-    padding?: TSpacing;
-    margin?: TSpacing;
-    children?: React.ReactNode;
-};
 
 const Grid: React.FC<PGrid> = ({
     row = false,
     component = 'div',
-    alignment = 'start',
-    justify = 'start',
+    alignment = 'flex-start',
+    justify = 'flex-start',
     flex = 0,
     children,
     ...rest
@@ -44,6 +26,6 @@ const Grid: React.FC<PGrid> = ({
     );
 };
 
-export default Grid;
-
 export type { PGrid };
+
+export default Grid;
