+++ src/foundations/layout/index.ts
@@ -1,0 +1,5 @@
+import Grid from './Grid';
+
+export * from './Grid.types';
+
+export default Grid;
