+++ src/foundations/shape/Shape.stories.mdx
@@ -6,7 +6,7 @@ import Shape from './index';
 import Grid from '../layout/Grid';
 import ThemeProvider from '../theme-provider/theme-provider';
 import Typography from '../typography';
-import spacing from '../layout/Spacing';
+import spacing from '../layout/Grid.spacing';
 
 export const Template = (args) => <Shape {...args} />;
 
@@ -73,8 +73,6 @@ There are 4 core shapes used throughout the Mattermost interface:
         </Grid>
     </Grid>
 </Grid>
-
-<ArgsTable of={Shape} />
 
 ## Rectangle
 
@@ -175,6 +173,8 @@ This shape is used in a few select places—namely the Mention Badge.
 The circle shape is mostly reserved for the Avatar component.
 This helps distinguish users from other UI elements throughout the product.
 There are a few other elements that use the circle shape, like the Unread Badge or the Status Badge.
+
+<ArgsTable of={Shape} />
 
 export const canvasExampleArgs = {
     ...defaultArgs,
