+++ src/foundations/typography/Typography.stories.mdx
@@ -5,7 +5,7 @@ import Typography from './Typography';
 import Grid from '../layout/Grid';
 import Shape from '../shape';
 import ThemeProvider from '../theme-provider/theme-provider';
-import spacing from '../layout/Spacing';
+import spacing from '../layout/Grid.spacing';
 
 <ThemeProvider />
